requirements.txt
APScheduler==3.7.0
black==20.8b1
bson==0.5.10
flake8==3.8.4
hmmlearn==0.2.5
ipykernel==5.5.0
Keras==2.4.3
mongoengine==0.22.1
mplfinance==0.12.7a7
newspaper3k==0.2.8
nsepy==0.8
pandas-market-calendars==1.6.1
pipdeptree==2.0.0
pmdarima==1.8.0
psycopg2-binary==2.8.6
PyMySQL==1.0.2
PyQt5==5.15.3
pytest @ file:///home/sarthak7gupta/.cache/pypoetry/artifacts/9d/40/6b/50d2cb06df545acea43c65379c5456f4fbc703115a703bc7cd433ddf70/pytest-5.4.3-py3-none-any.whl
seaborn==0.11.1
seqlearn==0.2
sklearn==0.0
SQLAlchemy-Utils==0.36.8
stockstats==0.3.2
tensorflow==2.4.1
textblob==0.15.3
